+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to [EffCSS](https://github.com/msabitov/effcss) are document
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [5.5.1] - 2026-09-04
+
+### Added
+- added tests for using utils within the scope
+
+### Fixed
+- fixed a bug with full ignoring falsy values ​​in variables
+- fixed a bug that did not reset the scope when there was a CSS generation error
+
+### Changed
+- layer tests have been refined
+
 ## [5.5.0] - 2026-08-28
 
 ### Added
@@ -787,7 +799,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ---
 
-[5.5.0]: https://github.com/msabitov/effcss
+[5.5.1]: https://github.com/msabitov/effcss
+[5.5.0]: https://github.com/msabitov/effcss/commit/0fa63d07e90163701e7b00c977abbdc58c6fffa5
 [5.4.5]: https://github.com/msabitov/effcss/commit/232a0cfb6bd55dfe4b6efed6aabc259d652890cd
 [5.4.4]: https://github.com/msabitov/effcss/commit/a15f952f02d5c972a1a63846eb6fd1ba87f18e7e
 [5.4.3]: https://github.com/msabitov/effcss/commit/ea9a1248abf7bb7711cef9379fd51fa1caada3c2